Update and rename README.md to DESCRIPTION.md
Description of base version that repository will start with. 
What remains to be investigated is finding an appropriate way of measuring the heat flux from the type-1 chain to the type-2 chain, and investigating effective ways of thermostatting the system so that the temperature of the type-1 chain remains fixed.