The edits made today were made in an attempt to identify where the issues with the increasing total energy are coming from.
The most significant changes were:
-To reduce the timestep of the integration from 0.005 to 0.001 (doesn't help)
-To try adding/removing a compute for potential energy (doesn't affect total energy computation)
-To try increasing the (initialisation) temperature, i.e. the Langevin temperature by one order of magnitude (energy still rises, but the difference in starting energies of chains is increased by factor of 10 as expected)

Not part of the commmit but present in local folder:
-A new input file and position file (name is identical save for removal of 'd='), where only one chain is present. Energy still appears to increase with time.

Also ran the original in.Ana file and found energy appears to decrease to a plataeu. This might mean the interaction of atoms with the walls is responsible; or the lack of a bond between the 1st and 16th atoms.