Nano input and position files
Input files with 'nano' in their name indicate files written using the nano unit
 style of lammps. This particular input file utilises thermostatting on only the
 first chain at a temperature of 298K. The second chain is shown to have rising
kinetic energy, but it is very, very low.

The dated pos file is one generated from the python script, pos_generator.py (to be included in next commit). Note that it has a mistake: the gap between chains is incorrectly specified to be equal to the atomic spacing. <THIS NEEDS TO BE CORRECTED>