New Python_scripts folder
A new folder has been added to the main directory for lammps work, called
Python_scripts. In this commit, it contains the pos_generator.py script
that can be used to create position files appropriate for the nano unit style
of lammps. Note, it still has a mistake, where the vacuum gap is not correctly
inserted into the pos file.

The other file was made by another lammps user, which reads lammps log files
and can create plots of the data using matplotlib.