pos_generator 2D update
pos_generator.py has been updated to include a method writing a 2D array of
atoms. Check the data dump folder for files dated 01-08 for examples. The latest
one is the 'most correct' (as of this commit, it is thought to be completely
correct as far as placing atoms is concerned). The .xyz file generated did not
show anything happening though, CHECK THIS.

The new 2D input file has also been included.

Older 1D input and pos files had also been modified today, probably to change
the type-1 type-2 relationship to a harmonic bond from LJ to check energy transf
er is higher with harmonic bond.