Double thermostatting with output
The input file has been modified so that both chains are separately Langevin thermostatted. Accordingly, the thermodynamical report has been modified to include 'f_' keywords in order to print *what is thought to be* the energy supplied by the Langevin thermostats to the chains.

However, the values are unexpected. The thermostat energy supplied to chain 1 is negative mostly, while positive energy is supplied to chain 2. It should be the other way round.

Also, note how the fix phonon and other related commands (by related I mean commands that share the same ID) have been commented out, showing they are not essential to the dynamics of the system.

Putting a harmonic bond instead of the LJ potential seems to give more intuitive results for the energy supplied and the KEs of the chains.