Adding PE compute
Have added a compute command that finds the total potential energy of the two chains. As of this commit, while the scripts seems to run fine (with the usual WARNING messages), the values still do not make much sense. The PE for chain-1 is positive while it is negative for chain-2.