2D input file in metal units
This commit contains 2 files: an updated .py file to generate 2D lattices (the functionality is the same, the 'update' refers to supplying different arguments to the function to generate today's files. The only actual change made was to the z-limits written onto the generated .pos file.

The 2D input file is in metal units. This is also the first variant, WHEN USING  A HAMAKER CONSTANT 50 TIMES GREATER THAN 1, to show the reservoirs' energy values having opposite signs: negative for the hot reservoir, and positive for the cold reservoir, as expected. What remains to be investigated, however, is the fact that the values are not of the same magnitude.